+++ CHANGES.md
@@ -56,10 +56,16 @@ To be released.
      -  The `--allow-private-address` or `-p` option allows looking up
         WebFinger information for private addresses (e.g., `localhost`).
 
+  -  Fixed a bug where the `fedify node` command had failed to correctly
+     render the favicon in terminal emulators that do not support 24-bit
+     colors.  [[#168], [#282] by Hyeonseo Kim]
+
+[#168]: https://github.com/fedify-dev/fedify/issues/168
 [#248]: https://github.com/fedify-dev/fedify/issues/248
 [#260]: https://github.com/fedify-dev/fedify/issues/260
 [#278]: https://github.com/fedify-dev/fedify/pull/278
 [#281]: https://github.com/fedify-dev/fedify/pull/281
+[#282]: https://github.com/fedify-dev/fedify/pull/282
 
 
 Version 1.7.3