+++ CHANGES.md
@@ -6,7 +6,7 @@ Fedify changelog
 Version 1.7.4
 -------------
 
-To be released.
+Released on July 13, 2025.
 
   - Fixed a bug the `-T`/`--no-tunnel` option in the `fedify inbox` command
     was being ignored, causing the server to always create a public tunnel